Renamed `tui_logic.py` to `utils.py` and refactored `ErrorScreen` to `modal_dialogs.py`
