+++ datashuttle/tui/custom_widgets.py
@@ -1,35 +1,7 @@
-from textual.app import ComposeResult
-from textual.containers import Container
 from textual.reactive import reactive
-from textual.screen import ModalScreen
-from textual.widgets import Button, Checkbox, Static
+from textual.widgets import Checkbox, Static
 
 from datashuttle.configs.canonical_configs import get_datatypes
-
-
-class ErrorScreen(ModalScreen):
-    """
-    A screen for rendering error messages. The border of the
-    central widget is red. The screen does not return any value.
-    """
-
-    def __init__(self, message):
-        super(ErrorScreen, self).__init__()
-
-        self.message = message
-
-    def compose(self) -> ComposeResult:
-        yield Container(
-            Container(
-                Static(self.message, id="errorscreen_message_label"),
-                id="errorscreen_message_container",
-            ),
-            Container(Button("OK"), id="errorscreen_ok_button"),
-            id="errorscreen_top_container",
-        )
-
-    def on_button_pressed(self) -> None:
-        self.dismiss()
 
 
 class DatatypeCheckboxes(Static):
