+++ datashuttle/tui/project_config.py
@@ -14,7 +14,7 @@ from textual.widgets import (
 )
 
 from datashuttle import DataShuttle
-from datashuttle.tui import tui_logic
+from datashuttle.tui.utils import utils
 
 
 class ShowConfigsDialog(ModalScreen):
@@ -268,9 +268,7 @@ class ConfigsContent(Container):
 
             self.parent_class.mainwindow.push_screen(
                 ShowConfigsDialog(
-                    tui_logic.get_textual_compatible_project_configs(
-                        project.cfg
-                    ),
+                    utils.get_textual_compatible_project_configs(project.cfg),
                     "A DataShuttle project with the below "
                     "configs has now been created.\n\n Click 'OK' to proceed to "
                     "the project page, where you will \n be able to create and "
@@ -294,7 +292,7 @@ class ConfigsContent(Container):
         try:
             self.project.make_config_file(**cfg_kwargs)
 
-            configs_to_show = tui_logic.get_textual_compatible_project_configs(
+            configs_to_show = utils.get_textual_compatible_project_configs(
                 self.project.cfg
             )
 
@@ -318,7 +316,7 @@ class ConfigsContent(Container):
         "ssh" widgets are hidden / displayed based on the current setting,
         in `self.switch_ssh_widgets_display()`.
         """
-        cfg_to_load = tui_logic.get_textual_compatible_project_configs(
+        cfg_to_load = utils.get_textual_compatible_project_configs(
             self.project.cfg
         )
